requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12

# Vector DB
chromadb==0.5.5

# Embeddings
sentence-transformers==3.2.1
transformers==4.45.2
torch>=2.0.0
Pillow==10.4.0

# OCR and PDF (Windows-friendly)
pytesseract==0.3.13
pypdf==4.3.1
pypdfium2==4.30.0

# Utilities
numpy==1.26.4
scikit-learn==1.5.2
orjson==3.10.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
requests==2.31.0
